ref: move validator to another file
