+++ src/app/auth/validators/index.ts
@@ -1,0 +1,1 @@
+export * from './password-match.validator';
